Fix super-linter issues: copyright, line length, markdown formatting

+++ src/Utils.hpp
@@ -1,3 +1,4 @@
+//  Copyright 2026 smart-swimmingpool
 #pragma once
 
 /**
@@ -10,17 +11,18 @@ namespace Utils {
 /**
      * Check if enough time has elapsed since last measurement
      * Handles millis() overflow correctly
-     * 
-     * @param lastMeasurement The last measurement timestamp in milliseconds
+     *
+     * @param lastMeasurement Last measurement timestamp (milliseconds)
      * @param intervalSeconds The interval in seconds
      * @return true if enough time has elapsed
      */
-inline bool shouldMeasure(unsigned long lastMeasurement, unsigned long intervalSeconds) {
+inline bool shouldMeasure(uint32_t lastMeasurement,
+                          uint32_t intervalSeconds) {
   if (lastMeasurement == 0) {
     return true;  // First measurement
   }
-  unsigned long currentMillis  = millis();
-  unsigned long intervalMillis = intervalSeconds * 1000UL;
+  uint32_t currentMillis  = millis();
+  uint32_t intervalMillis = intervalSeconds * 1000UL;
 
   // This handles overflow correctly
   return (currentMillis - lastMeasurement) >= intervalMillis;
@@ -28,14 +30,16 @@ inline bool shouldMeasure(unsigned long lastMeasurement, unsigned long intervalS
 
 /**
      * Convert float to string buffer with minimal heap usage
-     * 
+     *
      * @param value The float value to convert
-     * @param buffer The buffer to write to (minimum 16 bytes recommended)
+     * @param buffer The buffer to write to (min 16 bytes recommended)
      * @param bufferSize Size of the buffer (must be at least 8 bytes)
      * @param decimals Number of decimal places (default: 2)
-     * @note For typical temperature values (-50 to 100), 16 bytes is sufficient
+     * @note For typical temperature values (-50 to 100),
+     *       16 bytes is sufficient
      */
-inline void floatToString(float value, char* buffer, size_t bufferSize, int decimals = 2) {
+inline void floatToString(float value, char* buffer,
+                          size_t bufferSize, int decimals = 2) {
   // dtostrf needs minimum buffer size to avoid overflow
   if (bufferSize < 8) {
     buffer[0] = '\0';
@@ -46,7 +50,7 @@ inline void floatToString(float value, char* buffer, size_t bufferSize, int deci
 
 /**
      * Convert int to string buffer with minimal heap usage
-     * 
+     *
      * @param value The int value to convert
      * @param buffer The buffer to write to
      * @param bufferSize Size of the buffer
